requirements.txt
fastapi[standard]
uvicorn
psycopg[binary]
redis
slowapi
python-dotenv